Update backend operations
