--- brainpy/connectivity/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-from .base import *
-from .methods import *
